final_project/tests/requirements.txt
pytest
requests
sqla-wrapper
SQLAlchemy
pymysql
selenium
webdriver-manager
allure-pytest
pytest-xdist